Strategy/Template method refactor
