+++ dev/joymeter-webapi/src/main/java/com/joymeter/exception/ErrorCode.java
@@ -5,7 +5,8 @@ public enum ErrorCode {
 	INVALID_SESSION(98, "Invalid Session"),
 	UNAUTHORIZED(99, "Unauthorized"),
 	INVALID_FACEBOOK_TOKEN(100,"Invalid Facebook Access token"),
-	EXPIRED_FACEBOOK_TOKEN(101,"Expired Facebook Access token");
+	EXPIRED_FACEBOOK_TOKEN(101,"Expired Facebook Access token"),
+	INVALID_ACTIVITY(102,"Invalid activity id");
 
 	private long errorCode;
 	private String errorDescription;
