+++ dev/joymeter-webapi/src/main/java/com/joymeter/service/RecomendationService.java
@@ -1,0 +1,9 @@
+package com.joymeter.service;
+
+import com.joymeter.entity.Activity;
+
+public interface RecomendationService {
+	
+	Activity suggestActivity(long userId);
+
+}
