+++ dev/joymeter-webapi/src/main/resources/applicationContext.xml
@@ -16,7 +16,7 @@
 			">
 	<!--  Scan for both Jersey Rest Annotations a -->
 	<context:component-scan
-		base-package="com.joymeter.resource,com.joymeter.service,com.joymeter.service.impl,com.joymeter.service.jpa,com.joymeter.entity.util" />
+		base-package="com.joymeter.resource,com.joymeter.service,com.joymeter.service.impl,com.joymeter.repository,com.joymeter.repository.jpa,com.joymeter.entity.util" />
 	<context:annotation-config />
 	<tx:annotation-driven/>
